Bump version of arduino/compile-sketches action in workflow
The "CI Compile Examples" GitHub Actions workflow uses the arduino/compile-sketches action to compile the project's
Arduino sketches that demonstrate the usage of the bundled libraries.

A new release has been made of the arduino/compile-sketches action which fixes a bug that was causing spurious failures
of the workflow runs. Previously, the workflow pinned the action to the version that had the bug so it is necessary to
bump the pin in order to get the fixed version of the action.